rename 컴포넌트 common으로 변경
